fix: add procedure keyword support for procedure pointer declarations (partial fix for #1782)
This commit adds "procedure" to the type keyword lists in the parser to enable
recognition of procedure pointer declarations like:
    procedure(interface_name), pointer :: ptr

Changes made:
- Added "procedure" to parser dispatcher type keyword case
- Added "procedure" to statement utilities type keyword list
- Added "procedure" to all parser modules that check for type keywords
- Added test case for procedure pointer preservation

Current status:
The parser now recognizes procedure as a type keyword, but procedure pointer
declarations are still not being emitted in the output. Further investigation
needed to determine why declarations are being filtered or not generated.

Test shows that the declaration node appears to be created (since variable
is recognized), but the declaration itself is missing from output.

Refs #1782